add command `EndTxnOnPartition` implementation
